Updated Version of Electron App
Does not include node_modules folder as it was too big to include on GitHub. 

If you want to run this electron app, use the CLI to enter into the directory this code is cloned into and do these commands:

npm install electron --save-dev (installs electron and creates node-modules folder. --save-dev saves electron as a development dependency)
npm start (starts the application)